+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added bare-bones "%wasm": allows compiling cell to WASM and running that in the notebook. One
   can write widgets like this. Experimental: there are some use cases are not 100% clear. See
   "%help" for details on how this works.
+* Improved logging of errors; pre-checking for duplicate `package`, with improved error message.
 
 ## 0.8.0 -- Tests and Benchmarks, 2023/08/24